refactor(unused): rename the unused.py module as report.py, this is where all values calcualted for reporting purposes will go
